Split Credentials per ItemTypes
